8か月ぶりなので開発環境を最新に更新（CMake + VScodeのCubeIDE環境）

+++ src/STM32CubeIDE/Core/Src/sysmem.c
@@ -23,6 +23,7 @@
 /* Includes */
 #include <errno.h>
 #include <stdint.h>
+#include <stdlib.h>
 
 /**
  * Pointer to the current high watermark of the heap usage
@@ -50,7 +51,7 @@ static uint8_t *__sbrk_heap_end = NULL;
  * @param incr Memory size
  * @return Pointer to allocated memory
  */
-void *_sbrk(ptrdiff_t incr)
+void *_sbrk(int incr)
 {
   extern uint8_t _end; /* Symbol defined in the linker script */
   extern uint8_t _estack; /* Symbol defined in the linker script */
